% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd2ipynb.R
\name{html_nb_to_ipynb}
\alias{html_nb_to_ipynb}
\title{Convert an RMarkdown HTML notebook to an IPython (or Jupyter) notebook.}
\usage{
html_nb_to_ipynb(
  html_notebook_file,
  ipynb_file = str_replace(html_notebook_file, "\\\\.nb\\\\.html$", ".ipynb"),
  verbose = F,
  remove_html_comments = T
)
}
\arguments{
\item{html_notebook_file}{path to the RMarkdown HTML document to be converted}

\item{ipynb_file}{path to the desired output IPython notebook to be generated}
}
\value{
the resulting JSON of the IPython notebook, invisibly, as a character
vector
}
\description{
Note that you must pass the HTML file with extension \code{.nb.html} that is rendered
automatically for RMarkdown documents with output with \code{html_notebook} output.
Passing the RMarkdown document directly will fail.
}
